Remove map and table from country plan
