--- src/views/CountryThreeWNationalSocietyProjects/Filters/i18n.json
@@ -1,11 +1,0 @@
-{
-    "namespace": "countryThreeWNationalSocietyProjects",
-    "strings": {
-        "threeWFilterReportingNs": "Receiving Countries",
-        "threeWFilterProgrammeTypes": "Programme Types",
-        "threeWFilterSectors": "Sectors",
-        "threeWFilterTags": "Tags",
-        "threeWFilterProvinces": "Provinces",
-        "threeWFilterOperationTypes": "Project Types"
-    }
-}
