--- src/views/CountryThreeWNationalSocietyProjects/Filters/styles.module.css
@@ -1,5 +1,0 @@
-.filters {
-    display: grid;
-    grid-gap: var(--go-ui-spacing-md);
-    grid-template-columns: repeat(auto-fit, minmax(12rem, 1fr));
-}
