--- src/views/CountryThreeWNationalSocietyProjects/Map/i18n.json
@@ -1,8 +1,0 @@
-{
-    "namespace": "countryThreeWNationalSocietyProjects",
-    "strings": {
-        "reportingNationalSociety": "Reporting NS",
-        "receivingCountry": "Receiving Country",
-        "countryThreeWNSMapTitle": "{countryName} Projects"
-    }
-}
